Refactor PageHeader to hide implementation
