+++ web-ui/src/pages/_app.tsx
@@ -14,7 +14,7 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs'
 import { LocalizationProvider } from '@mui/x-date-pickers'
 import { LicenseInfo } from '@mui/x-license-pro'
-import { usePageHeader } from 'src/compositions/ui/pageTitle'
+import { usePageHeader } from 'src/compositions/ui/pageHeader'
 import 'dayjs/locale/en-gb'
 
 import 'react-perfect-scrollbar/dist/css/styles.css'
@@ -66,7 +66,7 @@ const App = (props: ExtendedAppProps) => {
 
   // Variables
   const getLayout = Component.getLayout ?? (page => <StandardVerticalLayout>{page}</StandardVerticalLayout>)
-  usePageHeader(s => s.setHeader)(<div></div>)
+  usePageHeader(s => s.setHeader)(<></>)
   return (
     <CacheProvider value={emotionCache}>
       <Head>
