+++ web-ui/src/pages/analytics/programs.tsx
@@ -2,12 +2,13 @@ import Grid from '@mui/material/Grid'
 import Typography from '@mui/material/Typography'
 import PageHeader from 'src/layouts/components/page-header'
 import TableSqlPrograms from 'src/analytics/TableSqlPrograms'
-import { usePageHeader } from 'src/compositions/ui/pageTitle'
+import { usePageHeader } from 'src/compositions/ui/pageHeader'
 
 const SqlPrograms = () => {
-  usePageHeader(s => s.setHeader)(
-    <PageHeader title='SQL Programs' subtitle='View status and edit already defined SQL programs.' />
-  )
+  usePageHeader(s => s.setHeader)({
+    title: 'SQL Programs',
+    subtitle: 'View status and edit already defined SQL programs.'
+  })
   return (
     <Grid container spacing={6} className='match-height'>
       <Grid item xs={12}>
