+++ web-ui/src/pages/connectors/list.tsx
@@ -3,10 +3,10 @@ import Typography from '@mui/material/Typography'
 
 import PageHeader from 'src/layouts/components/page-header'
 import DataSourceTable from 'src/connectors/DataSourceTable'
-import { usePageHeader } from 'src/compositions/ui/pageTitle'
+import { usePageHeader } from 'src/compositions/ui/pageHeader'
 
 const SqlPrograms = () => {
-  usePageHeader(s => s.setHeader)(<PageHeader title='Connectors' subtitle='View and edit data sources.' />)
+  usePageHeader(s => s.setHeader)({ title: 'Connectors', subtitle: 'View and edit data sources.' })
   return (
     <Grid container spacing={6} className='match-height'>
       <Grid item xs={12}>
