+++ web-ui/src/pages/streaming/management.tsx
@@ -2,14 +2,12 @@ import Grid from '@mui/material/Grid'
 import Typography from '@mui/material/Typography'
 import { ErrorBoundary } from 'react-error-boundary'
 import { ErrorOverlay } from 'src/components/table/ErrorOverlay'
-import { usePageHeader } from 'src/compositions/ui/pageTitle'
+import { usePageHeader } from 'src/compositions/ui/pageHeader'
 import PageHeader from 'src/layouts/components/page-header'
 import PipelineTable from 'src/streaming/management/PipelineTable'
 
 const PipelineManagement = () => {
-  usePageHeader(s => s.setHeader)(
-    <PageHeader title='Pipeline Management' subtitle='Start, stop and inspect pipelines.' />
-  )
+  usePageHeader(s => s.setHeader)({ title: 'Pipeline Management', subtitle: 'Start, stop and inspect pipelines.' })
   return (
     <Grid container spacing={6} className='match-height'>
       {/* id referenced by webui-tester */}
